feat: test and mock for patient

+++ test/models/patients.test.js
@@ -1,25 +1,25 @@
+require("dotenv").config();
 const db = require("../../src/models"); // Importe o objeto db completo
-const { Patient, Plan } = db; // Extraia os modelos do db
+const { Plan, Patient } = db; // Extraia os modelos do db
+
 
 describe("Patients Model", () => {
   beforeAll(async () => {
-    try {
-      // Desativa as verificações de FK
-      await db.sequelize.query("SET FOREIGN_KEY_CHECKS = 0");
-
-      await db.sequelize.sync({ force: true });
-
-      // Reativa as verificações
-      await db.sequelize.query("SET FOREIGN_KEY_CHECKS = 1");
-
-      await Plan.create({ coverage: "Basic", price: 100.0 });
-    } catch (error) {
-      console.error("Erro no beforeAll:", error);
+    if (process.env.NODE_ENV === 'test') {
+      // Criar o banco de dados de teste apenas no ambiente de teste
+      const database = process.env.DB_NAME || 'hospital_control_test';
+      try {
+        // Tente criar o banco de dados se ele não existir
+        await db.sequelize.query(`CREATE DATABASE IF NOT EXISTS ${database};`);
+        await db.sequelize.sync({ force: true }); // Sincronize as tabelas do banco de teste
+      } catch (error) {
+        console.error("Erro ao configurar o banco de dados de teste:", error);
+      }
     }
   });
 
   afterAll(async () => {
-    await db.sequelize.close();
+    await db.sequelize.close(); // Fecha a conexão com o banco após os testes
   });
 
   test("Deve criar um novo paciente", async () => {
